add pll module (atlpll), soc2 run at 25MHz
The ref clock is 50MHz, and the pll provides 25MHz.

The vga have to run at 25MHz to implement the 640x480 resolution.

Core, sram, and vga are all run at 25MHz now, beasue the sram and
vgaram bridge have to work on the same clock.

Later need a axi matrixi/bridge to separate sram and vgaram.

The system resetn should pull a longer time to make sure that the
pll is locked.

The pll doesnt have a reset pin, because cannot release the pll and
the reset of the soc at the same time.

Need a crg (clock & reset generator) module, to release reset for
the pll first, then use the refclk to count more cycle until the
pll is locked and then release reset for the reset of the soc.

Add test case soc2_test1_pll, the same thing as soc2_test0_vgaram,
just to observe the waves of clk and pll_clk_out_25Mhz.