amba_axi_m2s2 axi_interface axi_sram_bridge work together
amba_axi_m2s2 is generated from gen_amba project. It is a axi4
matrix with two masters and two slaves.

Did not try gen_amba_2021 axi matrix to see if it works, because
gen_amba_2021 version of amba_axi_m2s2 has not Mx_WID in the module
interfae, which I feel strange about.

On my AC415 FPGA board, need to press rst button to generate the
resetn signal to kick the whole thing start. The core runs fine even
without the reset, but the amba_axi_m2s2 need it. That's why there
is nothing on the screen earlier.

The axi_interface should give wlast=1 for non-burst transactions,
otherwise the amba_axi_m2s2 functions weird.

The problem left is that there are read transactions with
rresp=DECERR. Now axi_interface ignores it for a workaround.