change back to 75MHz; do not use inst_busy for controlling inst_req
Using inst_busy to control inst_req creates a combinational loop.
Inserting a register in between does not completely breaks the loop,
although it can be simulated in modelsim, it still impacts the design.

As a result of the inst_busy/inst_req loop, soc2 is unable to run at
75Mhz anymore, 50Mhz is still feasible

Additionally, this loop extends the systhesis time from 3 minutes to 8
minutes.